Added undeploy option to dev script

+++ dev.sh
@@ -3,6 +3,37 @@
 PLUGIN_NAME=cs-aware-connect
 CONTAINER_NAME=cs-connect-base
 
+# Options
+PRUNE_VOLUMES=false
+UNDEPLOY=false
+
+# Parse the command-line options
+while getopts "pu" opt; do
+    case $opt in
+        p)
+            echo "Prune volumes option set."
+            PRUNE_VOLUMES=true
+            ;;
+        u)
+            UNDEPLOY=true
+            ;;
+        \?)
+            echo "Invalid option: -$OPTARG"
+            exit 1
+            ;;
+    esac
+done
+
+if [ "$UNDEPLOY" = true ]; then
+    echo "Stopping containers..."
+    docker compose -f dev.docker-compose.yml down
+    echo "Containers stopped."
+    exit 0
+fi
+
+# Shift the parsed options, so $1 will point to the first non-option argument (if any).
+shift "$((OPTIND - 1))"
+
 echo "Stopping containers if running..."
 docker compose -f dev.docker-compose.yml down
 echo "Containers stopped."
@@ -29,6 +60,10 @@ echo "Starting containers..."
 docker compose -f dev.docker-compose.yml up -d
 echo "Containers started."
 
-echo "Cleaning up older volumes..."
-docker volume prune -f
+if [ "$PRUNE_VOLUMES" = true ]; then
+    echo "Cleaning up older volumes..."
+    docker volume prune -f
+    echo "Completed."
+fi
+
 echo "Completed."
